+++ rust/crates/photos/src/ml/mod.rs
@@ -16,4 +16,3 @@ pub mod runtime;
 pub mod types;
 #[cfg(feature = "usearch")]
 pub mod vector_db;
-mod webgpu;
