+++ rust/crates/photos/src/ml/onnx/webgpu.rs
